requirements.txt
openai
requests
python-dotenv
fastapi
uvicorn
pydantic
sendgrid
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
